services/rag_pipeline/requirements.txt
-r ../../requirements-shared.txt

# RAG Pipeline Specific
langchain-openai>=0.0.2
jieba>=0.42.1

# Document Loading
python-docx>=1.1.0
openpyxl>=3.1.2
pypdf>=3.17.0
# paddlepaddle>=2.6.0
# paddleocr>=2.7.0
